Add pledges/installments
Set some future Opportunities to be Stage=Pledged. For any that are
Pledged, create 4 Payments instead of only 1. Each of those 4 will have
a different Payment Date, spaced 4 months apart, with a Payment Amount
being 1/4 of the Opportunity Amount. If the Payment Date is past, mark
the Payment as Paid. Leave it unpaid otherwise. This finishes #107.

Also cleaned up some formatting inconsistencies and cleaned up some
remaining comments in the file.